refactor(channel): extract catalog model and lookup foundation
