+++ crates/app/src/channel/catalog/model.rs
@@ -1,0 +1,220 @@
+use serde::Serialize;
+
+use super::super::{ChannelCatalogTargetKind, ChannelPlatform};
+
+pub const CHANNEL_OPERATION_SEND_ID: &str = "send";
+pub const CHANNEL_OPERATION_SERVE_ID: &str = "serve";
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct ChannelRuntimeCommandDescriptor {
+    pub channel_id: &'static str,
+    pub platform: ChannelPlatform,
+    pub serve_bootstrap_agent_id: &'static str,
+}
+
+pub const TELEGRAM_RUNTIME_COMMAND_DESCRIPTOR: ChannelRuntimeCommandDescriptor =
+    ChannelRuntimeCommandDescriptor {
+        channel_id: "telegram",
+        platform: ChannelPlatform::Telegram,
+        serve_bootstrap_agent_id: "channel-telegram",
+    };
+
+pub const FEISHU_RUNTIME_COMMAND_DESCRIPTOR: ChannelRuntimeCommandDescriptor =
+    ChannelRuntimeCommandDescriptor {
+        channel_id: "feishu",
+        platform: ChannelPlatform::Feishu,
+        serve_bootstrap_agent_id: "channel-feishu",
+    };
+
+pub const MATRIX_RUNTIME_COMMAND_DESCRIPTOR: ChannelRuntimeCommandDescriptor =
+    ChannelRuntimeCommandDescriptor {
+        channel_id: "matrix",
+        platform: ChannelPlatform::Matrix,
+        serve_bootstrap_agent_id: "channel-matrix",
+    };
+
+pub const WECOM_RUNTIME_COMMAND_DESCRIPTOR: ChannelRuntimeCommandDescriptor =
+    ChannelRuntimeCommandDescriptor {
+        channel_id: "wecom",
+        platform: ChannelPlatform::Wecom,
+        serve_bootstrap_agent_id: "channel-wecom",
+    };
+
+pub const WHATSAPP_RUNTIME_COMMAND_DESCRIPTOR: ChannelRuntimeCommandDescriptor =
+    ChannelRuntimeCommandDescriptor {
+        channel_id: "whatsapp",
+        platform: ChannelPlatform::WhatsApp,
+        serve_bootstrap_agent_id: "channel-whatsapp",
+    };
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct ChannelCommandFamilyDescriptor {
+    pub runtime: ChannelRuntimeCommandDescriptor,
+    pub catalog: ChannelCatalogCommandFamilyDescriptor,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct ChannelCatalogCommandFamilyDescriptor {
+    pub channel_id: &'static str,
+    pub default_send_target_kind: ChannelCatalogTargetKind,
+    pub send: ChannelCatalogOperation,
+    pub serve: ChannelCatalogOperation,
+}
+
+impl ChannelCommandFamilyDescriptor {
+    pub fn channel_id(self) -> &'static str {
+        self.catalog.channel_id
+    }
+
+    pub fn default_send_target_kind(self) -> ChannelCatalogTargetKind {
+        self.catalog.default_send_target_kind
+    }
+
+    pub fn send(self) -> ChannelCatalogOperation {
+        self.catalog.send
+    }
+
+    pub fn serve(self) -> ChannelCatalogOperation {
+        self.catalog.serve
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+pub struct ChannelCatalogOperation {
+    pub id: &'static str,
+    pub label: &'static str,
+    pub command: &'static str,
+    pub availability: ChannelCatalogOperationAvailability,
+    pub tracks_runtime: bool,
+    pub requirements: &'static [ChannelCatalogOperationRequirement],
+    pub supported_target_kinds: &'static [ChannelCatalogTargetKind],
+}
+
+impl ChannelCatalogOperation {
+    pub fn supports_target_kind(self, kind: ChannelCatalogTargetKind) -> bool {
+        self.supported_target_kinds.contains(&kind)
+    }
+
+    pub fn default_target_kind(self) -> Option<ChannelCatalogTargetKind> {
+        self.supported_target_kinds.first().copied()
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+pub struct ChannelCatalogOperationRequirement {
+    pub id: &'static str,
+    pub label: &'static str,
+    pub config_paths: &'static [&'static str],
+    pub env_pointer_paths: &'static [&'static str],
+    pub default_env_var: Option<&'static str>,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+#[serde(rename_all = "snake_case")]
+pub enum ChannelCatalogOperationAvailability {
+    Implemented,
+    Stub,
+}
+
+impl ChannelCatalogOperationAvailability {
+    pub fn as_str(self) -> &'static str {
+        match self {
+            Self::Implemented => "implemented",
+            Self::Stub => "stub",
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+#[serde(rename_all = "snake_case")]
+pub enum ChannelCapability {
+    RuntimeBacked,
+    PluginBacked,
+    MultiAccount,
+    Send,
+    Serve,
+    RuntimeTracking,
+}
+
+impl ChannelCapability {
+    pub fn as_str(self) -> &'static str {
+        match self {
+            Self::RuntimeBacked => "runtime_backed",
+            Self::PluginBacked => "plugin_backed",
+            Self::MultiAccount => "multi_account",
+            Self::Send => "send",
+            Self::Serve => "serve",
+            Self::RuntimeTracking => "runtime_tracking",
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+#[serde(rename_all = "snake_case")]
+pub enum ChannelOnboardingStrategy {
+    ManualConfig,
+    PluginBridge,
+    Planned,
+}
+
+impl ChannelOnboardingStrategy {
+    pub fn as_str(self) -> &'static str {
+        match self {
+            Self::ManualConfig => "manual_config",
+            Self::PluginBridge => "plugin_bridge",
+            Self::Planned => "planned",
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+pub struct ChannelOnboardingDescriptor {
+    pub strategy: ChannelOnboardingStrategy,
+    pub setup_hint: &'static str,
+    pub status_command: &'static str,
+    pub repair_command: Option<&'static str>,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum ChannelDoctorCheckTrigger {
+    OperationHealth,
+    ReadyRuntime,
+    PluginBridgeHealth,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct ChannelDoctorCheckSpec {
+    pub name: &'static str,
+    pub trigger: ChannelDoctorCheckTrigger,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct ChannelDoctorOperationSpec {
+    pub checks: &'static [ChannelDoctorCheckSpec],
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct ChannelOperationDescriptor {
+    pub operation: ChannelCatalogOperation,
+    pub doctor: Option<ChannelDoctorOperationSpec>,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+#[serde(rename_all = "snake_case")]
+pub enum ChannelCatalogImplementationStatus {
+    RuntimeBacked,
+    ConfigBacked,
+    PluginBacked,
+    Stub,
+}
+
+impl ChannelCatalogImplementationStatus {
+    pub fn as_str(self) -> &'static str {
+        match self {
+            Self::RuntimeBacked => "runtime_backed",
+            Self::ConfigBacked => "config_backed",
+            Self::PluginBacked => "plugin_backed",
+            Self::Stub => "stub",
+        }
+    }
+}
